ime: include ImePurpose in IME platform output
We want to include purpose hints in IME output, so that android
keyboards know that they are inputting multiline or password. The
softkeyboard behaves differently depending on the purpose.

For example, on multiline the keyboards enter button changes from
the done action to a newline.

For passwords, suggestions are turned off, etc.

Signed-off-by: William Casarin <jb55@jb55.com>